first push

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: aaouassa <aaouassa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 13:26:40 by aaouassa          #+#    #+#             */
-/*   Updated: 2022/10/24 00:44:39 by aaouassa         ###   ########.fr       */
+/*   Updated: 2022/10/27 17:55:50 by aaouassa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char				*a;
 	const char			*b;
-	unsigned int		i;
+	size_t				i;
 
 	i = 0;
 	a = (char *)dst;
